Moving website code into header file
