+++ src/Main.cpp
@@ -7,11 +7,7 @@
 //All library includes
 #include <Arduino.h>
 #include <Wire.h>
-#include <WiFi.h>
 #include <YAAJ_ModbusMaster.h>
-#include <WiFiManager.h>
-#include <SPIFFS.h>
-#include <ESPAsyncWebServer.h>
 
 //all file includes
 #include "BioreactorVaribiles.hpp"
@@ -19,11 +15,7 @@
 #include "sensirion-lf.h"
 #include "sensirion-lf.cpp"
 #include "ESP_FlexyStepper.cpp"
-
-//Setup Wifi Manager and Server
-WiFiManager wifiManager;
-AsyncWebServer server(80);
-AsyncWebSocket ws("/ws");
+#include "WebHosting.hpp"
 
 //Set up Pump controller
 HardwareSerial ModbusSerial(1);
@@ -48,13 +40,6 @@ const float DISTANCE_PER_REV = 1;
 const float MAX_SPEED = 1000;
 const float ACCELERATION = 1000;
 const float MOVE_DISTANCE = 5; 
-
-//Declare Functions
-void initSPIFFS();
-void initWebServer();
-void initWebServer();
-void initWebSocket();
-void onEvent(AsyncWebSocket *server, AsyncWebSocketClient *client, AwsEventType type, void *arg, uint8_t *data, size_t len);
 
 //Start Running
 void setup() {
@@ -139,43 +124,4 @@ void loop() {
       //Do Nothing
   }
   Serial.print("Moved Stepper Motor\n");
-}
-
-//set up to host website
-//https://m1cr0lab-esp32.github.io/remote-control-with-websocket/web-server-setup/
-void initSPIFFS() {
-  if (!SPIFFS.begin()) {
-    Serial.println("Cannot mount SPIFFS volume...");
-  }
-}
-
-void onRootRequest(AsyncWebServerRequest *request) {
-  request->send(SPIFFS, "/index.html", "text/html", false);
-}
-
-void initWebServer() {
-  server.on("/", onRootRequest);
-  server.serveStatic("/", SPIFFS, "/");
-  server.begin();
-}
-
-void initWebSocket() {
-  ws.onEvent(onEvent);
-  server.addHandler(&ws);
-}
-
-void onEvent(AsyncWebSocket *server, AsyncWebSocketClient *client, AwsEventType type, void *arg, uint8_t *data, size_t len) {
-
-  switch (type) {
-    case WS_EVT_CONNECT:
-        Serial.printf("WebSocket client #%u connected from %s\n", client->id(), client->remoteIP().toString().c_str());
-        break;
-    case WS_EVT_DISCONNECT:
-        Serial.printf("WebSocket client #%u disconnected\n", client->id());
-        break;
-    case WS_EVT_DATA:
-    case WS_EVT_PONG:
-    case WS_EVT_ERROR:
-        break;
-  }
 }
